make RandomTetrominoGenerator abstract
